+++ src/TabList.js
@@ -14,7 +14,6 @@ export default class TabList extends React.Component {
 
 TabList.propTypes = {
   children: PropTypes.node.isRequired,
-  manager: PropTypes.object.isRequired,
   className: PropTypes.string,
   id: PropTypes.string,
   tag: PropTypes.string,