+++ src/index.js
@@ -1,0 +1,11 @@
+import Tab from './Tab';
+import TabList from './TabList';
+import TabPanel from './TabPanel';
+import Wrapper from './Wrapper';
+
+export default {
+  Tab,
+  TabList,
+  TabPanel,
+  Wrapper,
+};