+++ test/TabList-test.jsx
@@ -1,22 +1,22 @@
 import test from 'tape';
-import React from 'react/addons';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-addons-test-utils';
 import TabList from '../src/TabList';
-
-const ReactTestUtils = React.addons.TestUtils;
 
 test('TabList creation with only required props', t => {
   const element = ReactTestUtils.renderIntoDocument(
-    <TabList manager={{}}>
+    <TabList>
       <div>foo</div>
     </TabList>
   );
-  const node = React.findDOMNode(element);
+  const node = ReactDOM.findDOMNode(element);
 
-  t.equal(node.tagName, 'DIV');
+  t.equal(node.tagName.toLowerCase(), 'div');
   t.notOk(node.getAttribute('id'));
   t.notOk(node.getAttribute('class'));
   t.equal(node.getAttribute('role'), 'tablist');
-  t.equal(node.firstChild.tagName, 'DIV');
+  t.equal(node.firstChild.tagName.toLowerCase(), 'div');
   t.equal(node.firstChild.innerHTML, 'foo');
 
   t.end();
@@ -25,7 +25,6 @@ test('TabList creation with only required props', t => {
 test('TabList creation with all props', t => {
   const element = ReactTestUtils.renderIntoDocument(
     <TabList
-      manager={{}}
       id='bar'
       className='baz'
       tag='ul'
@@ -34,16 +33,16 @@ test('TabList creation with all props', t => {
       <p>foofoo</p>
     </TabList>
   );
-  const node = React.findDOMNode(element);
+  const node = ReactDOM.findDOMNode(element);
 
-  t.equal(node.tagName, 'UL');
+  t.equal(node.tagName.toLowerCase(), 'ul');
   t.equal(node.getAttribute('id'), 'bar');
   t.equal(node.getAttribute('class'), 'baz');
   t.equal(node.getAttribute('role'), 'tablist');
   t.equal(node.children.length, 2);
-  t.equal(node.firstChild.tagName, 'P');
+  t.equal(node.firstChild.tagName.toLowerCase(), 'p');
   t.equal(node.firstChild.innerHTML, 'foo');
-  t.equal(node.children[1].tagName, 'P');
+  t.equal(node.children[1].tagName.toLowerCase(), 'p');
   t.equal(node.children[1].innerHTML, 'foofoo');
 
   t.end();