+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Head
+- Upgrade to react 0.14 and its companion react-dom.
+
 ## 1.0.1
 - Fix PropTypes validation of TabPanel.